+++ src/components/organizations/show/members.js
@@ -5,6 +5,7 @@ import Icon from 'react-fa'
 import e from 'gEngine/engine'
 
 import './members.css'
+import {MemberAddForm} from '../shared/MemberAddForm/index.js'
 
 export class MembersTab extends Component {
   state = {
@@ -12,7 +13,7 @@ export class MembersTab extends Component {
   }
 
   render() {
-    const {members, memberships, invitations, admin_id, onRemove, addUser, httpRequests, meIsAdmin} = this.props
+    const {members, memberships, invitations, admin_id, onRemove, addUser, httpRequests, meIsAdmin, organizationId} = this.props
     const unjoinedInvitees = invitations.filter(i => !_.some(memberships, m => m.invitation_id === i.id))
     const onChangeSubTab = () => {this.setState({indexTabOpen: !this.state.indexTabOpen})}
     return (
@@ -21,7 +22,7 @@ export class MembersTab extends Component {
           <MembersIndexSubTab {...{members, invitations: unjoinedInvitees, admin_id, onRemove, onChangeSubTab, meIsAdmin}}/>
         }
         {!this.state.indexTabOpen &&
-          <MembersAddSubTab {...{addUser, httpRequests, onChangeSubTab}}/>
+          <MembersAddSubTab {...{organizationId, onChangeSubTab}}/>
         }
       </div>
     )
@@ -111,109 +112,17 @@ const Member = ({user, isAdmin, onRemove, meIsAdmin}) => (
   </div>
 )
 
-function validateEmail(email) {
-    var re = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
-    return re.test(email);
-}
-
-export class MembersAddSubTab extends Component{
-  state = { value: '' }
-
-  componentDidMount() {
-    this.refs.input.focus()
-  }
-
-  _submit() {
-    this.props.addUser(this.state.value)
-    this.setState({value: ''})
-  }
-
-  _onChange(e) {
-    this.setState({value: e.target.value})
-  }
-
-  _onKeyDown(e) {
-    if (e.keyCode === 13 && validateEmail(this.state.value)) {
-      this._submit();
-    }
-  }
-
-  render() {
-    const {value} = this.state
-    const isValid = validateEmail(value)
-    const isEmpty = _.isEmpty(value)
-    const buttonColor = (isValid || isEmpty) ? 'green' : 'grey'
-
-    const requests = _.orderBy(_.cloneDeep(this.props.httpRequests), ['created_at'], ['desc'])
-    return(
-      <div className='row MembersAddSubTab'>
-        <div className='col-sm-2'>
-          <div className='ui button large ' onClick={this.props.onChangeSubTab}>
-            <Icon name='chevron-left'/> Member List
-          </div>
-        </div>
-        <div className='col-sm-8'>
-          <h1> Invite New Members </h1>
-          <p> Members have viewing & editing access to all organization models. <br/> If you are on a plan, your pricing will be adjusted within 24 hours.</p>
-          <div className="ui form">
-            <div className="field">
-              <label>Email Address</label>
-              <input
-                value={this.state.value}
-                onKeyDown={this._onKeyDown.bind(this)}
-                type='text'
-                placeholder='name@domain.com'
-                ref='input'
-                onChange={this._onChange.bind(this)}
-              />
-            </div>
-            <div className={`ui button submit ${buttonColor} ${isValid ? '' : 'disabled'}`} onClick={this._submit.bind(this)}>
-              Invite User
-            </div>
-            <div>
-            </div>
-          </div>
-            {_.map(requests, (request) =>  {
-              return(
-                <InvitationHttpRequest
-                  key={request.id}
-                  busy={request.busy}
-                  success={request.success}
-                  email={request.email}
-                  isExistingMember={request.isExistingMember}
-                />
-              )
-            })}
-        </div>
+const MembersAddSubTab = ({organizationId, onChangeSubTab}) => (
+  <div className='row MembersAddSubTab'>
+    <div className='col-sm-2'>
+      <div className='ui button large ' onClick={onChangeSubTab}>
+        <Icon name='chevron-left'/> Member List
       </div>
-    )
-  }
-}
-
-const InvitationHttpRequest = ({busy, success, email, isExistingMember}) => {
-  let status = httpStatus(busy, success)
-
-  if (status === 'sending'){ return (
-    <div className='InvitationHttpRequest ui ignored message'>
-      Sending...
     </div>
-  )} else if (status === 'failure'){ return (
-    <div className='InvitationHttpRequest ui error message'>
-      Invitation to {email} failed.  This could be because they are already part of the organization or because of a server problem.  If it continues, please let us know.
+    <div className='col-sm-8'>
+      <h1> Invite New Members </h1>
+      <p> Members have viewing & editing access to all organization models. <br/> If you are on a plan, your pricing will be adjusted within 24 hours.</p>
+      <MemberAddForm organizationId={organizationId}/>
     </div>
-  )} else if (isExistingMember){ return (
-    <div className='InvitationHttpRequest ui success message'>
-      {email} was added to your organization.
-    </div>
-  )} else { return(
-    <div className='InvitationHttpRequest ui success message'>
-      {email} was sent an email invitation to join your organization.
-    </div>
-  )}
-}
-
-function httpStatus(busy, success) {
-  if (busy) { return 'sending' }
-  else if (success) { return 'success' }
-  else { return 'failure' }
-}
+  </div>
+)
